move ALL client classes to client source set
