+++ src/client/java/com/github/thedeathlycow/frostiful/client/anim/BiterAnimations.java
