+++ src/client/java/com/github/thedeathlycow/frostiful/client/mixin/SpriteAtlasManagerMixin.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.frostiful.mixins.client;
+package com.github.thedeathlycow.frostiful.client.mixin;
 
 import com.github.thedeathlycow.frostiful.Frostiful;
 import com.github.thedeathlycow.frostiful.client.registry.FTexturedRenderLayers;
