+++ src/client/java/com/github/thedeathlycow/frostiful/client/mixin/entity_renderer/RootedEffectRenderer.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.frostiful.mixins.client.entity_renderer;
+package com.github.thedeathlycow.frostiful.client.mixin.entity_renderer;
 
 import com.github.thedeathlycow.frostiful.entity.RootedEntity;
 import net.fabricmc.api.EnvType;
