+++ src/client/java/com/github/thedeathlycow/frostiful/client/model/BiterEntityModel.java
