+++ src/client/java/com/github/thedeathlycow/frostiful/client/model/FrostWandItemModel.java
