+++ src/client/java/com/github/thedeathlycow/frostiful/client/model/FrostologerEntityModel.java
