+++ src/client/java/com/github/thedeathlycow/frostiful/client/render/entity/BiterEntityRenderer.java
