+++ src/client/java/com/github/thedeathlycow/frostiful/client/render/entity/FrostologerEntityRenderer.java
