+++ src/client/java/com/github/thedeathlycow/frostiful/client/render/feature/CustomArmorTrimRenderer.java
