+++ src/client/java/com/github/thedeathlycow/frostiful/client/render/feature/FrostologerCloakFeatureRenderer.java
